refactor(web): share the optimistic notification mutation options
The archive, mark-seen and mark-all-seen option builders repeated the same
snapshot / rollback / invalidate wiring around a single differing optimistic
cache write.
