+++ apps/web/src/notifications/inbox/useMarkNotificationAsSeen.ts
@@ -9,13 +9,10 @@ import type {
 } from "@oboku/shared"
 import { type HttpApiClientWeb, useHttpClientApi } from "../../http"
 import { useActiveProfileId } from "../../profiles/active/activeProfileId"
+import { createNotificationMutationOptions } from "./createNotificationMutationOptions"
 import {
-  type NotificationCacheSnapshot,
-  cancelAndSnapshotNotificationQueries,
   inboxNotificationsQueryKey,
-  invalidateNotificationQueries,
   markSeenMutationKey,
-  rollbackNotificationCaches,
   unreadCountQueryKey,
 } from "./queryKeys"
 
@@ -23,41 +20,31 @@ export const markSeenMutationOptions = (
   queryClient: QueryClient,
   httpClientApi: HttpApiClientWeb,
   profileId: string | undefined,
-) => ({
-  mutationKey: markSeenMutationKey,
-  networkMode: "online" as const,
-  mutationFn: httpClientApi.markNotificationAsSeen,
-  onMutate: async ({ id }: { id: number }) => {
-    const snapshot = await cancelAndSnapshotNotificationQueries(
-      queryClient,
-      profileId,
-    )
-    const target = snapshot.previousNotifications?.find((n) => n.id === id)
+) =>
+  createNotificationMutationOptions({
+    queryClient,
+    profileId,
+    mutationKey: markSeenMutationKey,
+    mutationFn: httpClientApi.markNotificationAsSeen,
+    applyOptimisticUpdate: ({ id }: { id: number }, snapshot) => {
+      const target = snapshot.previousNotifications?.find((n) => n.id === id)
 
-    queryClient.setQueryData<GetNotificationsResponse>(
-      inboxNotificationsQueryKey(profileId),
-      (old) =>
-        old?.map((n) =>
-          n.id === id ? { ...n, seenAt: new Date().toISOString() } : n,
-        ),
-    )
-
-    if (target && !target.seenAt) {
-      queryClient.setQueryData<GetUnreadNotificationsCountResponse>(
-        unreadCountQueryKey(profileId),
-        (old) => (old ? { count: Math.max(0, old.count - 1) } : old),
+      queryClient.setQueryData<GetNotificationsResponse>(
+        inboxNotificationsQueryKey(profileId),
+        (old) =>
+          old?.map((n) =>
+            n.id === id ? { ...n, seenAt: new Date().toISOString() } : n,
+          ),
       )
-    }
 
-    return snapshot
-  },
-  onError: (
-    _err: unknown,
-    _vars: { id: number },
-    context: NotificationCacheSnapshot | undefined,
-  ) => rollbackNotificationCaches(queryClient, profileId, context),
-  onSettled: () => invalidateNotificationQueries(queryClient, profileId),
-})
+      if (target && !target.seenAt) {
+        queryClient.setQueryData<GetUnreadNotificationsCountResponse>(
+          unreadCountQueryKey(profileId),
+          (old) => (old ? { count: Math.max(0, old.count - 1) } : old),
+        )
+      }
+    },
+  })
 
 export const useMarkNotificationAsSeen = () => {
   const queryClient = useQueryClient()
